Functional data recorder, most sensors working
The Temperature+Humidity sensor, Wind sensor, CO2 sensor, SD card reader, Real time clock are all working independently of each other. The CO2 sensor has a conflict with the SD cards on pins 51 and 50, so it is impossible to record and read CO2 data at the same time. I have tested a version where we use the Uno to record and read CO2 data at the same time, so the problem is with the Mega... however the Uno solution only works if there is an extra 5V power supply from the Mega.